requirements.unified.txt
# MR.VERMA Unified - Comprehensive Dependencies
# Phase 1-2 Implementation Requirements
# Optimized for automated architecture generation

# ===========================================
# Core Web Framework
# ===========================================
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# ===========================================
# AI/ML APIs
# ===========================================
requests>=2.31.0
openai>=1.0.0

# ===========================================
# Data Processing
# ===========================================
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# ===========================================
# System Utilities
# ===========================================
psutil>=5.9.0
python-dotenv>=1.0.0

# ===========================================
# Terminal UI
# ===========================================
rich>=13.0.0
prompt-toolkit>=3.0.0

# ===========================================
# Data Validation & File Handling
# ===========================================
pydantic>=2.0.0
pyyaml>=6.0.0

# ===========================================
# Security
# ===========================================
cryptography>=41.0.0

# ===========================================
# Document Generation
# ===========================================
# Microsoft Office documents (Word, PowerPoint, Excel)
python-docx>=0.8.11       # Word documents
python-pptx>=0.6.21           # Create PowerPoint presentations
openpyxl>=3.1.2           # Excel spreadsheets
xlsxwriter>=3.1.9         # Excel alternative/fallback

# ===========================================
# Vector Embeddings & Semantic Search
# ===========================================
sentence-transformers>=2.2.2  # Embeddings generation
transformers>=4.30.0          # Hugging Face models
torch>=2.0.0                  # PyTorch backend
qdrant-client>=1.7.0          # Vector database client
faiss-cpu>=1.7.4              # Facebook AI Similarity Search (fallback)
chromadb>=0.4.0               # Alternative vector store

# ===========================================
# Database & ORM
# ===========================================
sqlalchemy>=2.0.0             # SQL ORM
alembic>=1.12.0               # Database migrations
pymongo>=4.6.0                # MongoDB driver
redis>=5.0.0                  # Redis cache/storage

# ===========================================
# Templating Engines
# ===========================================
jinja2>=3.1.2                 # Modern templating
mako>=1.2.4                   # Alternative templating

# ===========================================
# Optional Vector DB (commented by default)
# ===========================================
pymilvus>=2.3.0             # Milvus vector database (Primary)

# ===========================================
# External Tool Requirements (NOT installed via pip)
# ===========================================
# The following tools require external installation:
#
# Fluid (.NET):
#   - Download: https://github.com/sebastienros/fluid
#   - Requires: .NET SDK 6.0 or higher
#   - Installation: dotnet tool install --global fluid.cli
#
# Telosys (Java):
#   - Download: https://www.telosys.org/
#   - Requires: Java 11 or higher
#   - Installation: Manual download and PATH configuration
#
# Prisma (Node.js):
#   - Download: https://www.prisma.io/
#   - Requires: Node.js 16.x or higher
#   - Installation: npm install -g prisma
#
# Note: These tools are invoked via CLI from MR.VERMA
# and must be installed separately by the user.
